requirements.txt
customtkinter==5.2.1
mysql-connector-python==8.2.0
Pillow==10.1.0
reportlab==4.0.7
cloudinary==1.36.0
python-dotenv==1.0.0
requests==2.31.0
pywin32==306; platform_system == "Windows"
